fix: add code signing requirements to xpc connections
